deployment/requirements-vertex.txt
google-cloud-aiplatform[agent_engines,adk]>=1.40.0
google-cloud-secret-manager
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
aiohttp>=3.8.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
newspaper3k>=0.2.8
python-readability
requests>=2.28.0
asyncio>=3.4.3
typing-extensions>=4.0.0
pydantic>=2.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
